pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "flask>=3.1.2",
    "flask-dance>=7.1.0",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "google-genai>=1.53.0",
    "gunicorn>=23.0.0",
    "oauthlib>=3.3.1",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "requests>=2.32.5",
]
